small improvements as suggested by Paul on the dev list
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@1237 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/adapter/ApplicationAdapter.java
@@ -349,7 +349,7 @@ public class ApplicationAdapter extends StatefulScopeWrappingAdapter
 		try {
 			IStreamableFile streamFile = service.getStreamableFile(file);
 			ITagReader reader = streamFile.getReader();
-			duration = reader.getDuration() / 1000;
+			duration = (double) reader.getDuration() / 1000;
 			reader.close();
 		} catch (IOException e) {
 			log.error("error read stream file " + file.getAbsolutePath(), e);
